refactor(packs): move admin pack handlers to handlers.go
- Moved 5 admin handlers to handlers.go:
  - GetPacks (get all packs)
  - GetPackByID (get pack by ID)
  - PostPack (create pack)
  - PutPackByID (update pack)
  - DeletePackByID (delete pack)
- Preserved all swagger/OpenAPI documentation
- All handlers work identically
- Tests pass (46.1% coverage maintained)
